+++ sdk/php/src/StaticReferencesData.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types = 1);
+
 namespace AvtoDev\StaticReferencesData;
 
 use Exception;
@@ -8,23 +10,23 @@ use AvtoDev\StaticReferencesData\ReferencesData\StaticReference;
 class StaticReferencesData
 {
     /**
-     * Возвращает путь к корневой директории.
+     * Get root references directory path.
      *
      * @param string|null $additional_path
      *
      * @return string
      */
-    public static function getRootDirectoryPath($additional_path = null)
+    public static function getRootDirectoryPath(?string $additional_path = null): string
     {
-        $root = \dirname(\dirname(\dirname(__DIR__)));
+        $root = \dirname(__DIR__, 3);
 
-        return \is_string($additional_path) && ! empty($additional_path)
-            ? $root . DIRECTORY_SEPARATOR . ltrim((string) $additional_path, ' \\/')
+        return \is_string($additional_path)
+            ? $root . DIRECTORY_SEPARATOR . \ltrim($additional_path, ' \\/')
             : $root;
     }
 
     /**
-     * Returns static reference named 'auto categories'.
+     * Returns static reference named `auto categories`.
      *
      * @param string $file_name
      *
@@ -32,15 +34,13 @@ class StaticReferencesData
      *
      * @return StaticReference
      */
-    public static function getAutoCategories($file_name = 'auto_categories.json')
+    public static function getAutoCategories(string $file_name = 'auto_categories.json'): StaticReference
     {
-        return new StaticReference(
-            static::getRootDirectoryPath(sprintf('/data/auto_categories/%s', $file_name))
-        );
+        return new StaticReference(static::getRootDirectoryPath("/data/auto_categories/{$file_name}"));
     }
 
     /**
-     * Returns static reference named 'auto regions'.
+     * Returns static reference named `auto regions`.
      *
      * @param string $file_name
      *
@@ -48,15 +48,13 @@ class StaticReferencesData
      *
      * @return StaticReference
      */
-    public static function getAutoRegions($file_name = 'auto_regions.json')
+    public static function getAutoRegions(string $file_name = 'auto_regions.json'): StaticReference
     {
-        return new StaticReference(
-            static::getRootDirectoryPath(sprintf('/data/auto_regions/%s', $file_name))
-        );
+        return new StaticReference(static::getRootDirectoryPath("/data/auto_regions/{$file_name}"));
     }
 
     /**
-     * Returns static reference named 'registration actions'.
+     * Returns static reference named `registration actions`.
      *
      * @param string $file_name
      *
@@ -64,15 +62,13 @@ class StaticReferencesData
      *
      * @return StaticReference
      */
-    public static function getRegistrationActions($file_name = 'registration_actions.json')
+    public static function getRegistrationActions(string $file_name = 'registration_actions.json'): StaticReference
     {
-        return new StaticReference(
-            static::getRootDirectoryPath(sprintf('/data/registration_actions/%s', $file_name))
-        );
+        return new StaticReference(static::getRootDirectoryPath("/data/registration_actions/{$file_name}"));
     }
 
     /**
-     * Returns static reference named 'repair methods'.
+     * Returns static reference named `repair methods`.
      *
      * @param string $file_name
      *
@@ -80,15 +76,13 @@ class StaticReferencesData
      *
      * @return StaticReference
      */
-    public static function getRepairMethods($file_name = 'repair_methods.json')
+    public static function getRepairMethods(string $file_name = 'repair_methods.json'): StaticReference
     {
-        return new StaticReference(
-            static::getRootDirectoryPath(sprintf('/data/repair_methods/%s', $file_name))
-        );
+        return new StaticReference(static::getRootDirectoryPath("/data/repair_methods/{$file_name}"));
     }
 
     /**
-     * Returns static reference named 'auto fines'.
+     * Returns static reference named `auto fines`.
      *
      * @param string $file_name
      *
@@ -96,15 +90,13 @@ class StaticReferencesData
      *
      * @return StaticReference
      */
-    public static function getAutoFines($file_name = 'auto_fines.json')
+    public static function getAutoFines(string $file_name = 'auto_fines.json'): StaticReference
     {
-        return new StaticReference(
-            static::getRootDirectoryPath(sprintf('/data/auto_fines/%s', $file_name))
-        );
+        return new StaticReference(static::getRootDirectoryPath("/data/auto_fines/{$file_name}"));
     }
 
     /**
-     * Returns static reference named 'vehicle types'.
+     * Returns static reference named `vehicle types`.
      *
      * @param string $file_name
      *
@@ -112,15 +104,13 @@ class StaticReferencesData
      *
      * @return StaticReference
      */
-    public static function getVehicleTypes($file_name = 'vehicle_types.json')
+    public static function getVehicleTypes(string $file_name = 'vehicle_types.json'): StaticReference
     {
-        return new StaticReference(
-            static::getRootDirectoryPath(sprintf('/data/vehicle_types/%s', $file_name))
-        );
+        return new StaticReference(static::getRootDirectoryPath("/data/vehicle_types/{$file_name}"));
     }
 
     /**
-     * Returns static reference named 'cadastral districts'.
+     * Returns static reference named `cadastral districts`.
      *
      * @param string $file_name
      *
@@ -128,10 +118,8 @@ class StaticReferencesData
      *
      * @return StaticReference
      */
-    public static function getCadastralDistricts($file_name = 'cadastral_districts.json')
+    public static function getCadastralDistricts(string $file_name = 'cadastral_districts.json'): StaticReference
     {
-        return new StaticReference(
-            static::getRootDirectoryPath(sprintf('/data/cadastral_districts/%s', $file_name))
-        );
+        return new StaticReference(static::getRootDirectoryPath("/data/cadastral_districts/{$file_name}"));
     }
 }