+++ sdk/php/tests/ReferencesData/StaticReferenceTest.php
@@ -1,71 +1,108 @@
 <?php
 
+declare(strict_types = 1);
+
 namespace AvtoDev\StaticReferencesData\Tests\ReferencesData;
 
-use Exception;
+use InvalidArgumentException;
 use AvtoDev\StaticReferencesData\Tests\AbstractTestCase;
-use Tarampampam\Wrappers\Exceptions\JsonEncodeDecodeException;
 use AvtoDev\StaticReferencesData\ReferencesData\StaticReference;
 use AvtoDev\StaticReferencesData\ReferencesData\StaticReferenceInterface;
 
+/**
+ * @covers \AvtoDev\StaticReferencesData\ReferencesData\StaticReference<extended>
+ */
 class StaticReferenceTest extends AbstractTestCase
 {
     /**
-     * Test class implementations.
-     *
-     * @return void
+     * @var StaticReference
      */
-    public function testInstancesOf()
+    protected $reference;
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function setUp(): void
     {
-        $this->assertInstanceOf(StaticReferenceInterface::class, new StaticReference(__FILE__));
+        parent::setUp();
+
+        $this->reference = new StaticReference(__DIR__ . '/sample.json');
     }
 
     /**
-     * Test throwing exception on passing non-exists file into constructor.
-     *
      * @return void
      */
-    public function testThrowExceptionOnPassedNonExistsFile()
+    public function testInstancesOf(): void
     {
-        $this->expectException(Exception::class);
+        $this->assertInstanceOf(StaticReferenceInterface::class, $this->reference);
+    }
+
+    /**
+     * @return void
+     */
+    public function testThrowExceptionOnPassedNonExistsFile(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessageRegExp('~is not readable~i');
 
-        new StaticReference(null);
+        new StaticReference('');
     }
 
     /**
-     * Test throwing exception on passing invalid json-file file into constructor.
-     *
      * @return void
      */
-    public function testThrowExceptionOnPassedNonValidJson()
+    public function testGetHash(): void
     {
-        $this->expectException(JsonEncodeDecodeException::class);
-
-        $instance = new StaticReference(__DIR__ . '/../Stubs/wrong_json.json');
-        $instance->getContent();
+        $this->assertSame(\md5_file($this->reference->getFilePath(), false), $this->reference->getHash());
     }
 
     /**
-     * Test getters methods.
+     * @return void
+     */
+    public function testGetFilePath(): void
+    {
+        $this->assertSame(__DIR__ . '/sample.json', $this->reference->getFilePath());
+    }
+
+    /**
+     * @deprecated
      *
      * @return void
      */
-    public function testGetters()
+    public function testGetContent(): void
     {
-        $instance = new StaticReference(
-            $path = $this->getRootDirPath() . '/data/auto_categories/auto_categories.json'
-        );
+        $content = $this->reference->getContent();
 
-        $this->assertNotEmpty($hash = $instance->getHash());
-        $this->assertTrue(mb_strlen($hash) >= 8);
-        $this->assertRegExp('~[a-f0-9]~', $hash);
-        $this->assertEquals($hash, $instance->getHash());
+        $this->assertIsArray($content);
 
-        $this->assertEquals($path, $instance->getFilePath());
+        $this->assertArrayHasKey('foo', $content[0]);
+        $this->assertSame(1, $content[0]['foo']);
 
-        $this->assertInternalType('array', $content = $instance->getContent());
-        $this->assertNotEmpty($content);
-        $this->assertEquals($content, $instance->getContent());
+        $this->assertArrayHasKey('foo', $content[1]);
+        $this->assertSame(2, $content[1]['foo']);
+    }
+
+    /**
+     * @return void
+     */
+    public function testGetDataAsArray(): void
+    {
+        $as_array = $this->reference->getData(true);
+
+        $this->assertIsArray($as_array[0]);
+        $this->assertSame(1, $as_array[0]['foo']);
+        $this->assertSame(2, $as_array[1]['foo']);
+    }
+
+    /**
+     * @return void
+     */
+    public function testGetDataAsObject(): void
+    {
+        $as_object = $this->reference->getData(false);
+
+        $this->assertIsObject($as_object[0]);
+        $this->assertSame(1, $as_object[0]->foo);
+        $this->assertSame(2, $as_object[1]->foo);
     }
 }